refactor: webhook
- refactoring webhook event parsing
