+++ fuzz/fuzzers/webhook_push.rs
@@ -11,12 +11,12 @@ extern crate tiny_http;
 #[path = "../../src/metrics.rs"]
 mod metrics;
 
-#[path = "../../src/webhook.rs"]
-mod webhook;
+#[path = "../../src/webhook/event/push.rs"]
+mod push;
 
 fuzz_target!(|data: &[u8]| {
                  // fuzzed code goes here
                  if let Ok(s) = String::from_utf8(data.to_vec()) {
-                     webhook::handle_push(&s);
+                     let _ = push::Push::from_str(&s);
                  }
              });
